timer_test_read_config

+++ lab2/utils.c
@@ -6,7 +6,7 @@ int(util_get_LSB)(uint16_t val, uint8_t *lsb) {
   /* To be implemented by the students */
 
   if(lsb==NULL) return 1;
-  *lsb = val;
+  *lsb = (uint8_t) val;
   return 0;
 }
 
@@ -14,7 +14,7 @@ int(util_get_MSB)(uint16_t val, uint8_t *msb) {
   /* To be implemented by the students */
 
   if(msb==NULL) return 1;
-  *msb = val >> 8;
+  *msb = (uint8_t) (val >> 8);
   return 0;
 }
 
@@ -28,5 +28,6 @@ int (util_sys_inb)(int port, uint8_t *value) {
     return 1;
   }
   *value=new_val & 0xFF;
+  
   return 0;
 }
